pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ViTHash"
version = "0.1.0"
description = "A package to compute ViT-L/14-based image hashes (embeddings) and measure distance."
authors = [
    { name="Jelle De Geest", email="jelledegeest@hotmail.be" }
]
license = { text = "MIT" }
readme = "README.md"
keywords = ["clip", "vit", "image-hashing", "huggingface", "distance-metrics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
]

dependencies = [
    "torch>=1.9.0",
    "transformers>=4.21.0",
    "Pillow>=9.0.0",
    "numpy>=1.19.0"
]

[project.urls]
"Homepage" = "https://github.com/JelleDeGeest/ViTHash"